fix(notif): fixed notifications after big merge. notifications no longer use the store, as they did not persist anyway.
