+++ backend/src/messages/messages.controller.spec.ts
@@ -16,7 +16,6 @@ import { AchievementsLog } from 'src/achievements-log/entities/achievements-log.
 import { Match } from 'src/matches/entities/match.entity';
 import { NotificationsGateway } from 'src/notifications.gateway';
 import { ConfigService } from '@nestjs/config';
-import { Membership } from 'src/memberships/entities/membership.entity';
 
 describe('MessagesController', () => {
   let controller: MessagesController;
