+++ frontend/src/components/Notifications/NotificationIcon.vue
@@ -1,33 +1,29 @@
 <template>
-<v-menu
-  top="true"
-  right="true"
-  offset-x="true"
-  offset-overflow="true"
-  >
-  <template v-slot:activator="{ props }">
-  <v-btn icon
-        v-bind="props"
-        @click="toggle_alert"
-        >
-    <v-icon v-if="alert" color="accent">mdi-bell</v-icon>
-    <v-icon v-else>mdi-bell-outline</v-icon>
-  </v-btn>
+    <v-menu top="true" right="true" offset-x="true" offset-overflow="true">
+        <template v-slot:activator="{ props }">
+            <v-btn icon v-bind="props" @click="toggle_alert">
+                <v-icon v-if="alert" color="accent">mdi-bell</v-icon>
+                <v-icon v-else>mdi-bell-outline</v-icon>
+            </v-btn>
 
-</template>
-<v-list>
-    <v-list-item  v-for="item in notifications"
-                  :key="item.key"
-                  :to="item.route"
-                  :title="item.message">
-    </v-list-item>
-  </v-list>
-</v-menu>
+        </template>
+        <v-list>
+            <v-list-item v-for="item in notifications" :key="item.key" :to="item.route" :title="item.message">
+            </v-list-item>
+        </v-list>
+    </v-menu>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
 import { mapMutations } from 'vuex';
+
+interface Notification {
+  message: string;
+  userId: number;
+  key: number;
+  route: string;
+}
 
 export default defineComponent({
     data() {
@@ -35,7 +31,7 @@ export default defineComponent({
             alert: false,
             open: false,
             socket: this.$store.getters.socket,
-            notifications: this.$store.getters.notifications,
+            notifications: new Array<Notification>,
         };
     },
     created() {
@@ -43,17 +39,17 @@ export default defineComponent({
         this.socket.on("alert-challenge", (username: string, userId: number) => this.handleChallenge(username, userId));
     },
     methods: {
-      ...mapMutations(['pushNotification']),
+        ...mapMutations(['pushNotification']),
         toggle_alert() {
             this.alert = false;
         },
         handleChallenge(message: string, userId: number) {
             this.alert = true;
-            this.$store.dispatch('pushNotification', {message: message, userId: userId, route:`/game/`});
+            this.notifications.push({ message: message, userId: userId, route: `/game/`, key: this.notifications.length });
         },
         handleMessage(message: string) {
             this.alert = true;
-            this.$store.dispatch('pushNotification', {message: message, userId:0, route:"/chat"});
+            this.notifications.push({ message: message, userId: 0, route: "/chat", key: this.notifications.length });
         },
         reverseAlert() {
             this.alert = !this.alert;
